Move SubscribeTopicsCount related methods from MQTTProxyHandler to MQTTProxyProtocolMethodProcessor. (#125)
